config my-order

+++ src/Components/OrderCard/index.jsx
@@ -1,26 +1,36 @@
-import { XMarkIcon } from "@heroicons/react/24/solid"
+import { XMarkIcon } from "@heroicons/react/24/solid";
 
 const OrderCard = (props) => {
-  const {id, title, imageUrl, price, handleDelete } = props
+  const { id, title, imageUrl, price, handleDelete } = props;
+  let renderXMarkIcon;
+
+  if (handleDelete) {
+    renderXMarkIcon = (
+      <XMarkIcon
+        onClick={() => handleDelete(id)}
+        className="h-6 w-6 text-black cursor-pointer"
+      />
+    );
+  }
 
   return (
     <div className="flex justify-between items-center mb-2">
       <div className="flex items-center gap-2">
         <figure className="w-20 h-20">
-          <img className="w-full h-full rounded-lg object-cover" src={imageUrl} alt={title} />
+          <img
+            className="w-full h-full rounded-lg object-cover"
+            src={imageUrl}
+            alt={title}
+          />
         </figure>
-        <p className="text-sm font-light">
-          {title}
-        </p>
+        <p className="text-sm font-light">{title}</p>
       </div>
       <div className="flex items-center gap-2">
         <p className="text-lg font-medium">{price}</p>
-        <XMarkIcon 
-        onClick={() => handleDelete(id)}
-        className="h-6 w-6 text-black cursor-pointer" />
+        {renderXMarkIcon}
       </div>
     </div>
-  )
-}
+  );
+};
 
-export default OrderCard
+export default OrderCard;
